requirements.txt
# AlgoPortfolio Dependencies

# Telegram Bot
python-telegram-bot==20.7

# Stock Data
yfinance>=1.1.0

# Environment & Config
python-dotenv==1.0.0

# Async Support
aiohttp==3.9.1
aiosqlite>=0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock>=3.12.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0

# Type Checking
mypy==1.8.0
